+++ Source/sharp_pare_0x15D8.hpp
@@ -91,7 +91,7 @@ struct sharp_pare_0x15D8
     EXPORT s16 sub_5B9220(u16 a2, u16 a3);
     EXPORT void ReadTextures_5B92E0();
     EXPORT void sub_5B9350();
-    EXPORT STexture* sub_5B94F0(s32 type, s32 pal, s32 kind, s32 a5);
+    EXPORT STexture* sub_5B94F0(s32 type, u16 pal, s32 kind, s32 a5);
     EXPORT STexture* GetTexture2_5B95D0(u16 textureIdx);
     EXPORT STexture* sub_5B95F0(u16 a2, u16 a3, u16 a4);
     EXPORT void sub_5B9660(u16 texture_idx, u16 pal_idx);
